Fix (clippy) warnings

+++ src/values.rs
@@ -17,12 +17,12 @@ impl TryFrom<crate::csgo_proto::c_msg_source1_legacy_game_event::KeyT> for RawVa
             1 if value.val_string.is_some() => Ok(Self::String(value.val_string.unwrap())),
             2 if value.val_float.is_some() => Ok(Self::F32(value.val_float.unwrap())),
             3 if value.val_long.is_some() => Ok(Self::I32(value.val_long.unwrap())),
-            4 if value.val_short.is_some() => Ok(Self::I32(value.val_short.unwrap() as i32)),
-            5 if value.val_byte.is_some() => Ok(Self::I32(value.val_byte.unwrap() as i32)),
+            4 if value.val_short.is_some() => Ok(Self::I32(value.val_short.unwrap())),
+            5 if value.val_byte.is_some() => Ok(Self::I32(value.val_byte.unwrap())),
             6 if value.val_bool.is_some() => Ok(Self::Bool(value.val_bool.unwrap())),
             7 if value.val_uint64.is_some() => Ok(Self::U64(value.val_uint64.unwrap())),
             8 if value.val_long.is_some() => Ok(Self::I32(value.val_long.unwrap())),
-            9 if value.val_short.is_some() => Ok(Self::I32(value.val_short.unwrap() as i32)),
+            9 if value.val_short.is_some() => Ok(Self::I32(value.val_short.unwrap())),
             _ => Err(()),
         }
     }
